Fix mwc and mqb conversion warnings and introduce safeCast

+++ src/groups/mwc/mwcc/mwcc_orderedhashmap.t.cpp
@@ -145,7 +145,7 @@ static void test1_breathingTest()
     bsl::pair<IterType, bool> rc = map.insert(bsl::make_pair(1, s));
     ASSERT_EQ(true, rc.first != map.end());
     ASSERT_EQ(rc.second, true);
-    ASSERT_EQ(1, rc.first->first);
+    ASSERT_EQ(1U, rc.first->first);
     ASSERT_EQ(s, rc.first->second);
     ASSERT_EQ(1U, cmap.count(1));
 
@@ -286,8 +286,8 @@ static void test4_rinsert()
         RcType rc = map.rinsert(bsl::make_pair(i, i + 1));
         ASSERT_EQ_D(i, true, rc.second);
         ASSERT_EQ_D(i, true, rc.first != map.end());
-        ASSERT_EQ_D(i, i, rc.first->first);
-        ASSERT_EQ_D(i, (i + 1), rc.first->second);
+        ASSERT_EQ_D(i, static_cast<int>(i), rc.first->first);
+        ASSERT_EQ_D(i, static_cast<int>(i + 1), rc.first->second);
         ASSERT_EQ_D(i, true, 1.5 >= map.load_factor());
     }
 
@@ -298,8 +298,8 @@ static void test4_rinsert()
         const MyMapType& cmap = map;
         size_t           i    = k_NUM_ELEMENTS - 1;
         for (ConstIterType cit = cmap.begin(); cit != cmap.end(); ++cit) {
-            ASSERT_EQ_D(i, i, cit->first);
-            ASSERT_EQ_D(i, (i + 1), cit->second);
+            ASSERT_EQ_D(i, static_cast<int>(i), cit->first);
+            ASSERT_EQ_D(i, static_cast<int>(i + 1), cit->second);
             --i;
         }
     }
@@ -311,13 +311,13 @@ static void test4_rinsert()
         ConstIterType    cit  = --(cmap.end());  // last element
         for (; cit != cmap.begin(); --cit) {
             ASSERT_EQ_D(i, true, i < k_NUM_ELEMENTS);
-            ASSERT_EQ_D(i, i, cit->first);
-            ASSERT_EQ_D(i, (i + 1), cit->second);
+            ASSERT_EQ_D(i, static_cast<int>(i), cit->first);
+            ASSERT_EQ_D(i, static_cast<int>(i + 1), cit->second);
             ++i;
         }
         ASSERT_EQ(true, cit == cmap.begin());
-        ASSERT_EQ(i, cit->first);
-        ASSERT_EQ(cit->second, (i + 1));
+        ASSERT_EQ(static_cast<int>(i), cit->first);
+        ASSERT_EQ(cit->second, static_cast<int>(i + 1));
     }
 }
 
@@ -457,8 +457,8 @@ static void test6_clear()
         RcType rc = map.insert(bsl::make_pair(i, i + 1));
         ASSERT_EQ_D(i, rc.second, true);
         ASSERT_EQ_D(i, true, rc.first != map.end());
-        ASSERT_EQ_D(i, i, rc.first->first);
-        ASSERT_EQ_D(i, (i + 1), rc.first->second);
+        ASSERT_EQ_D(i, static_cast<int>(i), rc.first->first);
+        ASSERT_EQ_D(i, static_cast<int>(i + 1), rc.first->second);
     }
 
     ASSERT_EQ(false, map.empty());
@@ -493,8 +493,8 @@ static void test7_erase()
         RcType rc = map.insert(bsl::make_pair(i, i));
         ASSERT_EQ_D(i, rc.second, true);
         ASSERT_EQ_D(i, true, rc.first != map.end());
-        ASSERT_EQ_D(i, i, rc.first->first);
-        ASSERT_EQ_D(i, i, rc.first->second);
+        ASSERT_EQ_D(i, static_cast<int>(i), rc.first->first);
+        ASSERT_EQ_D(i, static_cast<int>(i), rc.first->second);
     }
 
     const MyMapType& cmap = map;
@@ -610,7 +610,7 @@ static void test10_erasureIterator()
     // Find
     IterType iter = map.find(9000);
     ASSERT_EQ(true, iter != map.end());
-    ASSERT_EQ(iter->first, 9000);
+    ASSERT_EQ(iter->first, 9000U);
 
     // erase
     IterType it = map.erase(iter);
@@ -768,27 +768,27 @@ static void test13_previousEndIterator()
     ASSERT_EQ(true, rc.first == endIt);
     ASSERT_EQ(true, rc.first == endCit);
 
-    ASSERT_EQ(i, endIt->first);
-    ASSERT_EQ(i * i, endIt->second);
+    ASSERT_EQ(static_cast<int>(i), endIt->first);
+    ASSERT_EQ(static_cast<int>(i * i), endIt->second);
 
     ++i;
     for (; i < 10000; ++i) {
         endIt = map.end();
         rc    = map.insert(bsl::make_pair(i, i * i));
         ASSERT_EQ_D(i, true, rc.first == endIt);
-        ASSERT_EQ_D(i, i, endIt->first);
-        ASSERT_EQ_D(i, i * i, endIt->second);
+        ASSERT_EQ_D(i, static_cast<int>(i), endIt->first);
+        ASSERT_EQ_D(i, static_cast<int>(i * i), endIt->second);
     }
 
     // Erase last element
-    map.erase(i - 1);
-    ASSERT_EQ((i - 2), (--map.end())->first);
+    map.erase(static_cast<int>(i - 1));
+    ASSERT_EQ(static_cast<int>(i - 2), (--map.end())->first);
     endIt = map.end();
     ++i;
     rc = map.insert(bsl::make_pair(i, i * i));
     ASSERT_EQ(true, rc.first == endIt);
-    ASSERT_EQ(i, endIt->first);
-    ASSERT_EQ(i * i, endIt->second);
+    ASSERT_EQ(static_cast<int>(i), endIt->first);
+    ASSERT_EQ(static_cast<int>(i * i), endIt->second);
 
     // rinsert an element, which doesn't affect end().
     ++i;
@@ -798,8 +798,8 @@ static void test13_previousEndIterator()
     ++i;
     rc = map.insert(bsl::make_pair(i, i * i));
     ASSERT_EQ(true, endIt == rc.first);
-    ASSERT_EQ(i, endIt->first);
-    ASSERT_EQ(i * i, endIt->second);
+    ASSERT_EQ(static_cast<int>(i), endIt->first);
+    ASSERT_EQ(static_cast<int>(i * i), endIt->second);
 }
 
 static void test14_localIterator()
@@ -917,8 +917,8 @@ static void test15_eraseRange()
     ASSERT(map.erase(map.begin(), second) == second);
     ASSERT(map.begin() == second);
     ASSERT_EQ(k_NUM_ELEMENTS - 1, map.size());
-    ASSERT_EQ_D(1, 1, map.begin()->first);
-    ASSERT_EQ_D(1, 1, map.begin()->second);
+    ASSERT_EQ_D(1, 1U, map.begin()->first);
+    ASSERT_EQ_D(1, 1U, map.begin()->second);
 
     ASSERT(map.erase(--map.end(), map.end()) == map.end());
     ASSERT_EQ(k_NUM_ELEMENTS - 2, map.size());
@@ -1098,7 +1098,7 @@ testN1_insertPerformanceUnordered_GoogleBenchmark(benchmark::State& state)
         typedef bsl::unordered_map<size_t, size_t> MyMapType;
         MyMapType map(state.range(0), s_allocator_p);
         for (auto _ : state) {
-            for (size_t i = 0; i < state.range(0); ++i) {
+            for (size_t i = 0; i < static_cast<size_t>(state.range(0)); ++i) {
                 map.insert(bsl::make_pair(i, i));
             }
         }
@@ -1115,9 +1115,9 @@ testN1_insertPerformanceOrdered_GoogleBenchmark(benchmark::State& state)
         // OrderedHashMap
         typedef mwcc::OrderedHashMap<size_t, size_t> MyMapType;
 
-        MyMapType map(state.range(0), s_allocator_p);
+        MyMapType map(static_cast<int>(state.range(0)), s_allocator_p);
         for (auto _ : state) {
-            for (size_t i = 0; i < state.range(0); ++i) {
+            for (size_t i = 0; i < static_cast<size_t>(state.range(0)); ++i) {
                 map.insert(bsl::make_pair(i, i));
             }
         }
@@ -1135,7 +1135,7 @@ testN2_erasePerformanceUnordered_GoogleBenchmark(benchmark::State& state)
     MyMapType map(s_allocator_p);
     for (auto _ : state) {
         state.PauseTiming();
-        for (size_t i = 0; i < state.range(0); ++i) {
+        for (size_t i = 0; i < static_cast<size_t>(state.range(0)); ++i) {
             RcType rc = map.insert(bsl::make_pair(i, i));
             ASSERT_EQ_D(i, i, rc.first->first);
             ASSERT_EQ_D(i, i, rc.first->second);
@@ -1165,7 +1165,7 @@ testN2_erasePerformanceOrdered_GoogleBenchmark(benchmark::State& state)
     // Insert 1M elements
     for (auto _ : state) {
         state.PauseTiming();
-        for (size_t i = 0; i < state.range(0); ++i) {
+        for (size_t i = 0; i < static_cast<size_t>(state.range(0)); ++i) {
             RcType rc = map.insert(bsl::make_pair(i, i));
             ASSERT_EQ_D(i, i, rc.first->first);
             ASSERT_EQ_D(i, i, rc.first->second);
@@ -1190,11 +1190,11 @@ static void testN3_profile_GoogleBenchmark(benchmark::State& state)
     // This case can be used to profile the component.
 
     typedef mwcc::OrderedHashMap<size_t, size_t> MyMapType;
-    MyMapType map(state.range(0), s_allocator_p);
+    MyMapType map(static_cast<int>(state.range(0)), s_allocator_p);
 
     // Insert elements.
     for (auto _ : state) {
-        for (size_t i = 0; i < state.range(0); ++i) {
+        for (size_t i = 0; i < static_cast<size_t>(state.range(0)); ++i) {
             map.insert(bsl::make_pair(i, i));
         }
     }
